requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.54.0
slack-bolt==1.18.0
slack-sdk==3.26.0
schedule==1.2.0
